hotfix: 네일 그룹 / 세트 엔티티 Long -> Integer로 hotfix

+++ src/main/java/art/snail/naillian/backend/domain/nail/entity/NailGroup.java
@@ -12,11 +12,11 @@ import org.springframework.data.relational.core.mapping.Table;
 @NoArgsConstructor
 public class NailGroup {
     @Id
-    private Long id;
+    private Integer id;
 
-    private Long fingerThumb;
-    private Long fingerIndex;
-    private Long fingerMiddle;
-    private Long fingerRing;
-    private Long fingerPinky;
+    private Integer fingerThumb;
+    private Integer fingerIndex;
+    private Integer fingerMiddle;
+    private Integer fingerRing;
+    private Integer fingerPinky;
 }
